requirements.txt
asgiref==3.5.2
boto3==1.24.92
botocore==1.27.92
cachetools==5.2.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
da-vinci==0.3.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-pb-model==0.3.3
django-storages==1.13.1
django-thumbnails==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.4
igdb-api-v4==0.0.5
jmespath==1.0.1
oauthlib==3.2.1
Pillow==9.2.0
pipenv==2022.9.8
platformdirs==2.5.2
protobuf==4.21.7
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodomex==3.15.0
pydot==1.4.2
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
requests-oauthlib==1.3.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
shortuuid==1.0.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
vdf==3.4
